fix: step2 피드백 반영: 자료형변경 (double -> long)

+++ src/main/java/lotto/domain/Rank.java
@@ -24,7 +24,7 @@ public enum Rank {
                 .orElse(MISS);
     }
 
-    public double getWinningMoney() {
+    public long getWinningMoney() {
         return winningMoney;
     }
 }
